backend/requirements.txt
aenum==2.1.2
aiohttp==3.7.4.post0
alembic==1.6.5
aniso8601==9.0.1
argh==0.26.2
asgiref==3.3.4
asn1crypto==1.4.0
async-generator==1.10
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==21.2.0
Authlib==0.10
autopep8==1.6.0
azure-common==1.1.27
azure-core==1.15.0
azure-storage-blob==12.8.1
Babel==2.9.1
backcall==0.2.0
bcrypt==3.2.0
blinker==1.4
boto3==1.17.95
botocore==1.20.95
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2
click-default-group==1.2.2
cloudinary==1.25.0
colorama==0.4.4
cryptography==3.4.7
decorator==4.4.2
distlib==0.3.4
Django==3.2.1
django-cors-headers==3.11.0
django-filter==21.1
django-rest-auth==0.9.5
djangorestframework==3.13.1
dnspython==2.1.0
email-validator==1.1.3
fasteners==0.15
filelock==3.6.0
Flask==1.1.4
Flask-BabelEx==0.9.4
Flask-Cors==3.0.10
Flask-Executor==0.9.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
flatten-dict==0.1.0
greenlet==1.1.0
gunicorn==19.10.0
idna==2.10
insultgenerator==1.0.3
ipython==7.24.1
ipython-genutils==0.2.0
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
jmespath==0.10.0
jsonschema==2.6.0
loremipsum==1.0.5
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
meltano==1.67.0
meltano-flask-security==0.1.0
monotonic==1.6
msrest==0.6.21
multidict==5.1.0
networkx==2.5.1
numpy==1.22.1
oauthlib==3.1.1
opencv-python==4.5.5.62
oscrypto==1.2.1
parso==0.8.2
passlib==1.7.4
pathlib2==2.3.5
pathtools==0.1.2
pickleshare==0.7.5
Pillow==8.2.0
platformdirs==2.5.1
prompt-toolkit==3.0.18
psutil==5.8.0
psycopg2==2.9.3
psycopg2-binary==2.8.6
pycodestyle==2.8.0
pycparser==2.20
pycryptodomex==3.10.1
Pygments==2.9.0
pyhocon==0.3.57
pyhumps==1.2.2
PyJWT==2.1.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyPika==0.25.2
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
python-gitlab==1.15.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.4.2
simplejson==3.17.2
six==1.16.0
smtpapi==0.4.7
snowflake-connector-python==2.4.5
snowflake-sqlalchemy==1.2.4
speaklater==1.3
SQLAlchemy==1.4.18
sqlparse==0.3.1
toml==0.10.2
traitlets==5.0.5
typing-extensions==3.10.0.0
tzdata==2021.5
urllib3==1.26.5
virtualenv==20.13.4
waitress==2.0.0
watchdog==0.9.0
wcwidth==0.2.5
Werkzeug==0.16.1
White-Noise==0.1.0
whitenoise==5.2.0
WTForms==2.3.3
yarl==1.6.3
